[ch20055] Update Siesta for .NET 4.72
This change is a bit of a hack, but it's basically about partially
undoing the insanity that is dependency injection. I could have taken
the cleanup farther, but opted to just hack in enough to get it
working.

The basic change here is to update RestModelBinder to do the
serialization itself rather than relying on a bunch of DI magic to
figure out types. And it turns out that somewhere along the way,
something stopped supporting the bonkers forms-encoded JSON thing that
we do, so I had to re-implement it.